refactor: update useLoanInfo
