+++ src/pages/Loans/LoansOverview/components/BorrowLimit/BorrowLimit.tsx
@@ -38,7 +38,10 @@ const BorrowLimit = ({
   console.log('ray : ***** currentBorrowLimit?.toString() => ', currentBorrowLimit?.toString());
   // ray test touch >>
 
+  // ray test touch <<
+  // TODO: update with `calculateBorrowLimitBtcChange`
   const newBorrowLimit = getBorrowLimitUSD({ type: loanAction, amount: actionAmount, asset });
+  // ray test touch >>
   const newLTV = getLTV({ type: loanAction, amount: actionAmount, asset });
 
   if (!currentLTV || !newLTV || !currentBorrowLimit || !newBorrowLimit) {
