+++ src/pages/Loans/LoansOverview/components/LoansInsights/LoansInsights.tsx
@@ -3,7 +3,7 @@ import { useMutation } from 'react-query';
 import { formatNumber, formatPercentage, formatUSD } from '@/common/utils/utils';
 import { Card, CTA, Dl, DlGroup } from '@/component-library';
 import ErrorModal from '@/legacy-components/ErrorModal';
-import { AccountPositionsStatisticsData } from '@/utils/hooks/api/loans/use-get-account-positions';
+import { AccountPositionsStatisticsData } from '@/utils/hooks/api/loans/lend-and-borrow-info';
 import { useGetAccountSubsidyRewards } from '@/utils/hooks/api/loans/use-get-account-subsidy-rewards';
 
 import { StyledDd, StyledDt } from './LoansInsights.style';
